meetandgreet/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
python-decouple==3.8
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==5.0.6
executing==2.0.1
ipython==8.24.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pytz==2024.1
pyyaml==6.0.1
uritemplate==4.1.1
setuptools==70.0.0
